third_party/WebKit/LayoutTests/fast/html/imports/import-script-expected.txt

Test that scripts in imported documents are executed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS labels is loadedLabels
PASS successfullyParsed is true

TEST COMPLETE
